+++ __tests__/unit/scales/base.spec.ts
@@ -1,4 +1,4 @@
-import { Base } from '../../../src';
+import { Base } from '../../../src/scales/base';
 import { BaseOptions, Domain, Range } from '../../../src/types';
 
 class Scale extends Base<BaseOptions> {
