+++ __tests__/unit/scales/continuous.spec.ts
@@ -1,5 +1,5 @@
 import { identity } from '@antv/util';
-import { Continuous } from '../../../src';
+import { Continuous } from '../../../src/scales/continuous';
 import { ContinuousOptions } from '../../../src/types';
 import { createInterpolate } from '../../../src/utils';
